requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
playwright==1.58.0
pyee==13.0.1
python-telegram-bot==21.3
requests==2.31.0
typing_extensions==4.15.0
urllib3==2.6.3
